requirements.txt
opensearch-py>=2.4.0
rich>=13.0.0
click>=8.1.0
urllib3>=1.26.0
simple-term-menu>=1.6.0
python-dotenv==1.2.2
psutil>=5.9.0